requirements.txt
openai
pandas
sqlalchemy
python-dotenv
numpy
scikit-learn
requests
Pillow
openai-agents